Moved some more methods of AbstractRandom to IEnhancedGenerator extension methods.  Reorganized location of IEnhancedRandom and related files.
